added missing const support

+++ CS2X.CoreLib/Single.cs
@@ -4,12 +4,12 @@ namespace System
 {
 	public readonly struct Single
 	{
-		public const float MinValue = (float)-3.40282346638528859e+38;
-        public const float Epsilon = (float)1.4e-45;
-        public const float MaxValue = (float)3.40282346638528859e+38;
-        public const float PositiveInfinity = (float)1.0 / (float)0.0;
-        public const float NegativeInfinity = (float)-1.0 / (float)0.0;
-        public const float NaN = (float)0.0 / (float)0.0;
+		public const float MinValue = -3.40282346638528859e+38f;
+        public const float Epsilon = 1.4e-45f;
+        public const float MaxValue = 3.40282346638528859e+38f;
+        public const float PositiveInfinity = 1.0f / 0.0f;
+        public const float NegativeInfinity = -1.0f / 0.0f;
+        public const float NaN = 0.0f / 0.0f;
 
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public static unsafe bool IsNaN(float f)
